4789: Rename ProjectStepMove to Timeline::StepMove
