+++ app/controllers/admin/timeline_step_moves_controller.rb
@@ -1,10 +1,10 @@
-class Admin::ProjectStepMovesController < Admin::AdminController
+class Admin::TimelineStepMovesController < Admin::AdminController
   include TranslationSaveable, LogControllable
 
   def new
     @step = ProjectStep.find(params[:step_id])
     authorize @step, :update?
-    @step_move = ProjectStepMove.new(
+    @step_move = Timeline::StepMove.new(
       step: @step,
       days_shifted: params[:days_shifted],
       context: params[:context]
@@ -20,7 +20,7 @@ class Admin::ProjectStepMovesController < Admin::AdminController
     authorize @step, :update?
     @log = ProjectLog.new(project_log_params)
     authorize @log, :create?
-    @step_move = ProjectStepMove.new(project_step_move_params.merge(step: @step, log: @log))
+    @step_move = Timeline::StepMove.new(project_step_move_params.merge(step: @step, log: @log))
 
     @step_move.execute!
     @log.save!
@@ -30,6 +30,6 @@ class Admin::ProjectStepMovesController < Admin::AdminController
   private
 
   def project_step_move_params
-    params.require(:project_step_move).permit(:move_type, :shift_subsequent, :days_shifted, :context)
+    params.require(:timeline_step_move).permit(:move_type, :shift_subsequent, :days_shifted, :context)
   end
 end
